Rename kernel vpp file to better reflect its function
