+++ src/libcharon/plugins/kernel_vpp/Makefile.am
@@ -16,7 +16,7 @@ endif
 libstrongswan_kernel_vpp_la_SOURCES = \
     kernel_vpp_plugin.h kernel_vpp_plugin.c \
     kernel_vpp_net.h kernel_vpp_net.c \
-    kernel_vpp_shared.h kernel_vpp_shared.c \
+    kernel_vpp_grpc.h kernel_vpp_grpc.c \
     # kernel_vpp_ipsec.h kernel_vpp_ipsec.c
 
 libstrongswan_kernel_vpp_la_LDFLAGS = -module -avoid-version \
